feat: Move catalog service out of `cli` package
