+++ cli/commands/catalog/tui/model.go
@@ -6,9 +6,9 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/catalog/service"
-	"github.com/gruntwork-io/terragrunt/cli/commands/catalog/service/module"
 	"github.com/gruntwork-io/terragrunt/cli/commands/catalog/tui/components/buttonbar"
+	"github.com/gruntwork-io/terragrunt/internal/services/catalog"
+	"github.com/gruntwork-io/terragrunt/internal/services/catalog/module"
 	"github.com/gruntwork-io/terragrunt/options"
 )
 
@@ -47,7 +47,7 @@ func (b button) String() string {
 type model struct {
 	list              list.Model
 	terragruntOptions *options.TerragruntOptions
-	svc               service.CatalogService
+	svc               catalog.CatalogService
 	selectedModule    *module.Module
 	delegateKeys      *delegateKeyMap
 	buttonBar         *buttonbar.ButtonBar
@@ -62,7 +62,7 @@ type model struct {
 	ready             bool
 }
 
-func newModel(opts *options.TerragruntOptions, svc service.CatalogService) model {
+func newModel(opts *options.TerragruntOptions, svc catalog.CatalogService) model {
 	var (
 		modules      = svc.Modules()
 		items        = make([]list.Item, 0, len(modules))
