+++ cli/commands/catalog/tui/tui.go
@@ -7,11 +7,11 @@ import (
 
 	tea "github.com/charmbracelet/bubbletea"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/catalog/service"
+	"github.com/gruntwork-io/terragrunt/internal/services/catalog"
 	"github.com/gruntwork-io/terragrunt/options"
 )
 
-func Run(ctx context.Context, opts *options.TerragruntOptions, svc service.CatalogService) error {
+func Run(ctx context.Context, opts *options.TerragruntOptions, svc catalog.CatalogService) error {
 	if _, err := tea.NewProgram(newModel(opts, svc), tea.WithAltScreen(), tea.WithContext(ctx)).Run(); err != nil {
 		if err := context.Cause(ctx); errors.Is(err, context.Canceled) {
 			return nil
