+++ internal/services/catalog/module/doc.go
