+++ internal/services/catalog/module/doc_test.go
