+++ internal/services/catalog/module/module.go
