+++ internal/services/catalog/module/testdata/find_modules/gitdir/HEAD
