+++ internal/services/catalog/module/testdata/find_modules/modules/eks-alb-ingress-controller-iam-policy/README.md
