+++ internal/services/catalog/module/testdata/find_modules/modules/eks-aws-auth-merger/README.adoc
